move TwigHelper into template package
