+++ src/fr/adrienbrault/idea/symfony2plugin/action/SymfonySymbolSearchAction.java
@@ -22,7 +22,7 @@ import com.intellij.util.indexing.IdFilter;
 import com.jetbrains.php.lang.psi.elements.PhpClass;
 import fr.adrienbrault.idea.symfony2plugin.Symfony2Icons;
 import fr.adrienbrault.idea.symfony2plugin.Symfony2ProjectComponent;
-import fr.adrienbrault.idea.symfony2plugin.TwigHelper;
+import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 import fr.adrienbrault.idea.symfony2plugin.action.model.SymfonySymbolSearchModel;
 import fr.adrienbrault.idea.symfony2plugin.dic.ContainerService;
 import fr.adrienbrault.idea.symfony2plugin.doctrine.EntityHelper;
