+++ src/fr/adrienbrault/idea/symfony2plugin/asset/AssetGoToDeclarationHandler.java
@@ -7,7 +7,7 @@ import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiManager;
 import fr.adrienbrault.idea.symfony2plugin.Symfony2ProjectComponent;
-import fr.adrienbrault.idea.symfony2plugin.TwigHelper;
+import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 import org.apache.commons.lang.ArrayUtils;
 import org.jetbrains.annotations.Nullable;
