+++ src/fr/adrienbrault/idea/symfony2plugin/stubs/indexes/TwigExtendsStubIndex.java
@@ -12,7 +12,7 @@ import com.jetbrains.twig.TwigFile;
 import com.jetbrains.twig.TwigFileType;
 import com.jetbrains.twig.elements.TwigExtendsTag;
 import fr.adrienbrault.idea.symfony2plugin.Symfony2ProjectComponent;
-import fr.adrienbrault.idea.symfony2plugin.TwigHelper;
+import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 import gnu.trove.THashMap;
 import org.jetbrains.annotations.NotNull;
