+++ src/fr/adrienbrault/idea/symfony2plugin/templating/TemplateReference.java
@@ -5,7 +5,7 @@ import com.intellij.psi.PsiElementResolveResult;
 import com.intellij.psi.PsiPolyVariantReferenceBase;
 import com.intellij.psi.ResolveResult;
 import com.jetbrains.php.lang.psi.elements.StringLiteralExpression;
-import fr.adrienbrault.idea.symfony2plugin.TwigHelper;
+import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
