+++ tests/fr/adrienbrault/idea/symfony2plugin/tests/templating/path/JsonFileIndexTwigNamespacesTest.java
@@ -2,7 +2,7 @@ package fr.adrienbrault.idea.symfony2plugin.tests.templating.path;
 
 import com.intellij.openapi.util.Condition;
 import com.intellij.util.containers.ContainerUtil;
-import fr.adrienbrault.idea.symfony2plugin.TwigHelper;
+import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 import fr.adrienbrault.idea.symfony2plugin.templating.path.TwigPath;
 import fr.adrienbrault.idea.symfony2plugin.templating.path.TwigPathIndex;
 import fr.adrienbrault.idea.symfony2plugin.tests.SymfonyLightCodeInsightFixtureTestCase;
