+++ tests/fr/adrienbrault/idea/symfony2plugin/tests/templating/translation/TwigTranslationCompletionContributorTest.java
@@ -1,6 +1,7 @@
 package fr.adrienbrault.idea.symfony2plugin.tests.templating.translation;
 
 import com.jetbrains.twig.TwigFileType;
+import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 
 /**
  * @author Daniel Espendiller <daniel@espendiller.net>
@@ -73,7 +74,7 @@ public class TwigTranslationCompletionContributorTest extends TwigTranslationFix
     }
 
     /**
-     * @see fr.adrienbrault.idea.symfony2plugin.TwigHelper#getTransDefaultDomainPattern
+     * @see TwigHelper#getTransDefaultDomainPattern
      * @see fr.adrienbrault.idea.symfony2plugin.templating.TwigTemplateCompletionContributor
      */
     public void testTwigTransDefaultDomainDomainCompletion() {
@@ -83,7 +84,7 @@ public class TwigTranslationCompletionContributorTest extends TwigTranslationFix
     }
 
     /**
-     * @see fr.adrienbrault.idea.symfony2plugin.TwigHelper#getTranslationTokenTagFromPattern
+     * @see TwigHelper#getTranslationTokenTagFromPattern
      */
     public void testTranslationTokenTagFromCompletion() {
         assertCompletionContains(TwigFileType.INSTANCE, "{% trans from \"<caret>\" %}", "interchange");
