+++ tests/fr/adrienbrault/idea/symfony2plugin/tests/templating/util/TwigHelperTempTest.java
@@ -1,13 +1,13 @@
-package fr.adrienbrault.idea.symfony2plugin.tests;
+package fr.adrienbrault.idea.symfony2plugin.tests.templating.util;
 
 import com.intellij.openapi.util.Pair;
 import com.intellij.psi.PsiDirectory;
-import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFile;
 import fr.adrienbrault.idea.symfony2plugin.Settings;
-import fr.adrienbrault.idea.symfony2plugin.TwigHelper;
+import fr.adrienbrault.idea.symfony2plugin.templating.util.TwigHelper;
 import fr.adrienbrault.idea.symfony2plugin.templating.path.TwigNamespaceSetting;
 import fr.adrienbrault.idea.symfony2plugin.templating.path.TwigPathIndex;
+import fr.adrienbrault.idea.symfony2plugin.tests.SymfonyTempCodeInsightFixtureTestCase;
 import fr.adrienbrault.idea.symfony2plugin.util.yaml.YamlPsiElementFactory;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.yaml.psi.YAMLFile;
@@ -15,16 +15,15 @@ import org.jetbrains.yaml.psi.YAMLFile;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
-import java.util.function.Predicate;
 
 /**
  * @author Daniel Espendiller <daniel@espendiller.net>
  *
- * @see fr.adrienbrault.idea.symfony2plugin.TwigHelper
+ * @see TwigHelper
  */
 public class TwigHelperTempTest extends SymfonyTempCodeInsightFixtureTestCase {
     /**
-     * @see fr.adrienbrault.idea.symfony2plugin.TwigHelper#getTemplatePsiElements
+     * @see TwigHelper#getTemplatePsiElements
      */
     public void testGetTemplatePsiElements() {
         createFile("res/foo.html.twig");
@@ -48,7 +47,7 @@ public class TwigHelperTempTest extends SymfonyTempCodeInsightFixtureTestCase {
     }
 
     /**
-     * @see fr.adrienbrault.idea.symfony2plugin.TwigHelper#getTemplateNamesForFile
+     * @see TwigHelper#getTemplateNamesForFile
      */
     public void testGetTemplateNamesForFile() {
         Settings.getInstance(getProject()).twigNamespaces.addAll(createTwigNamespaceSettings());
@@ -65,7 +64,7 @@ public class TwigHelperTempTest extends SymfonyTempCodeInsightFixtureTestCase {
     }
 
     /**
-     * @see fr.adrienbrault.idea.symfony2plugin.TwigHelper#getTwigFileNames
+     * @see TwigHelper#getTwigFileNames
      */
     public void testGetTwigFileNames() {
         createFile("res/foobar/foo.html.twig");
@@ -79,7 +78,7 @@ public class TwigHelperTempTest extends SymfonyTempCodeInsightFixtureTestCase {
     }
 
     /**
-     * @see fr.adrienbrault.idea.symfony2plugin.TwigHelper#getTwigAndPhpTemplateFiles
+     * @see TwigHelper#getTwigAndPhpTemplateFiles
      */
     public void testGetTwigAndPhpTemplateFiles() {
         createFiles("res/foobar/foo.html.twig", "res/foobar/foo.php");
@@ -94,7 +93,7 @@ public class TwigHelperTempTest extends SymfonyTempCodeInsightFixtureTestCase {
     }
 
     /**
-     * @see fr.adrienbrault.idea.symfony2plugin.TwigHelper#getTemplateNavigationOnOffset
+     * @see TwigHelper#getTemplateNavigationOnOffset
      */
     public void testGetTemplateNavigationOnOffset() {
         createFiles("res/foobar/foo.html.twig");
@@ -110,7 +109,7 @@ public class TwigHelperTempTest extends SymfonyTempCodeInsightFixtureTestCase {
     }
 
     /**
-     * @see fr.adrienbrault.idea.symfony2plugin.TwigHelper#getTemplateTargetOnOffset
+     * @see TwigHelper#getTemplateTargetOnOffset
      */
     public void testGetTemplateTargetOnOffset() {
         createFiles("res/foobar/foo.html.twig");
@@ -143,7 +142,7 @@ public class TwigHelperTempTest extends SymfonyTempCodeInsightFixtureTestCase {
     }
 
     /**
-     * @see fr.adrienbrault.idea.symfony2plugin.TwigHelper#getTwigPathFromYamlConfigResolved
+     * @see TwigHelper#getTwigPathFromYamlConfigResolved
      */
     public void testGetTwigPathFromYamlConfigResolved() {
         createFile("app/test/foo.yaml");
